pyproject.toml
[tool.poetry]
name = "fileserver"
version = "0.1.0"
description = ""
authors = ["Зимин Николай <Nikola.Zimin@x5.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
orjson = "^3.8.3"
sqlalchemy = "^1.4.45"
alembic = "^1.9.1"
asyncpg = "^0.27.0"
pytz = "^2022.7"
coloredlogs = "^15.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"
redis = "^4.4.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.20.3"
flake8 = "^6.0.0"
isort = "^5.11.4"
mypy = "^0.991"
flake8-isort = "^6.0.0"
flake8-commas = "^2.1.0"
httpx = "^0.23.1"
types-aiofiles = "^22.1.0.4"
types-redis = "^4.3.21.7"
types-pytz = "^2022.7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_mode = "auto"
python_classes = "*Test Test*"
python_files = "test_*.py *_test.py"
python_functions = "test_*  *_test"
norecursedirs = ".venv venv"
filterwarnings = ["ignore::DeprecationWarning"]
markers = ["slow: Show slow tests", "new: Run only new tests"]

[tool.mypy]
exclude = "src/db/migration/"
python_version = 3.11
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
implicit_optional = true
